refactor: Названия папок и файлов, роут 404
