+++ src/components/Preview/Preview.jsx
@@ -1,5 +1,5 @@
 import { useNavigate } from 'react-router-dom';
-import './styles.css';
+import './Preview.css';
 
 export const Preview = props => {
   const navigate = useNavigate();
