+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
-import { App } from './App/App';
+import { App } from './components/App/App';
 import './index.css';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
