+++ src/pages/404/404.css
