+++ src/pages/CatalogExperts/CatalogExperts.jsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
-import { Preview } from '../../components/Preview';
+import { Preview } from '../../components/Preview/Preview';
 import { useNavigate } from 'react-router-dom';
-import './styles.css';
+import './CatalogExperts.css';
 
 export const CatalogExecutors = () => {
   const navigate = useNavigate();
@@ -19,9 +19,7 @@ export const CatalogExecutors = () => {
       <button onClick={() => navigate(`/client/${1}`, { state: { name: 'Клиент 1', id: 1 } })}>
         Личный кабинет клиента
       </button>
-      <button
-        onClick={() => navigate(`/executor/${1}`, { state: { name: 'Исполнитель 1', id: 1 } })}
-      >
+      <button onClick={() => navigate(`/expert/${1}`, { state: { name: 'Исполнитель 1', id: 1 } })}>
         Личный кабинет исполнителя
       </button>
     </div>
