+++ src/pages/ClientAccount/ClientAccount.jsx
@@ -1,12 +1,12 @@
 import { useLocation, useNavigate } from 'react-router-dom';
-import './styles.css';
-export const ClientChat = () => {
+import './ClientAccount.css';
+export const ClientAccount = () => {
   const navigate = useNavigate();
   const { state } = useLocation();
   const { name, id } = state;
   return (
-    <div className={'clientChat-container'}>
-      <h1>Чат клиента: {name}</h1>
+    <div className={'clientAccount-container'}>
+      <h1>Личный кабинет клиента: {name}</h1>
       <button onClick={() => navigate('/catalog')}>Перейти в каталог</button>
       <button onClick={() => navigate(`/client/${id}/orders`, { state: { name, id } })}>
         Заказы
