+++ src/pages/ClientChat/ClientChat.jsx
@@ -1,14 +1,14 @@
 import { useLocation, useNavigate } from 'react-router-dom';
-import './styles.css';
-export const ExecutorAccount = () => {
+import './ClientChat.css';
+export const ClientChat = () => {
   const navigate = useNavigate();
   const { state } = useLocation();
   const { name, id } = state;
   return (
-    <div className={'executorAccount-container'}>
-      <h1>Личный кабинет исполнителя: {name}</h1>
+    <div className={'clientChat-container'}>
+      <h1>Чат клиента: {name}</h1>
       <button onClick={() => navigate('/catalog')}>Перейти в каталог</button>
-      <button onClick={() => navigate(`/executor/${id}/orders`, { state: { name, id } })}>
+      <button onClick={() => navigate(`/client/${id}/orders`, { state: { name, id } })}>
         Заказы
       </button>
     </div>
