+++ src/pages/ExpertAccount/ExpertAccount.jsx
@@ -1,14 +1,14 @@
 import { useLocation, useNavigate } from 'react-router-dom';
-import './styles.css';
-export const ExecutorChat = () => {
+import './ExpertAccount.css';
+export const ExecutorAccount = () => {
   const navigate = useNavigate();
   const { state } = useLocation();
   const { name, id } = state;
   return (
-    <div className={'executorChat-container'}>
-      <h1>Чат исполнителя: {name}</h1>
+    <div className={'expertAccount-container'}>
+      <h1>Личный кабинет исполнителя: {name}</h1>
       <button onClick={() => navigate('/catalog')}>Перейти в каталог</button>
-      <button onClick={() => navigate(`/executor/${id}/orders`, { state: { name, id } })}>
+      <button onClick={() => navigate(`/expert/${id}/orders`, { state: { name, id } })}>
         Заказы
       </button>
     </div>
