+++ src/pages/ExpertChat/ExpertChat.jsx
@@ -1,14 +1,14 @@
 import { useLocation, useNavigate } from 'react-router-dom';
-import './styles.css';
-export const ClientAccount = () => {
+import './ExpertChat.css';
+export const ExecutorChat = () => {
   const navigate = useNavigate();
   const { state } = useLocation();
   const { name, id } = state;
   return (
-    <div className={'clientAccount-container'}>
-      <h1>Личный кабинет клиента: {name}</h1>
+    <div className={'expertChat-container'}>
+      <h1>Чат исполнителя: {name}</h1>
       <button onClick={() => navigate('/catalog')}>Перейти в каталог</button>
-      <button onClick={() => navigate(`/client/${id}/orders`, { state: { name, id } })}>
+      <button onClick={() => navigate(`/expert/${id}/orders`, { state: { name, id } })}>
         Заказы
       </button>
     </div>
