+++ src/pages/ExpertRatings/ExpertRatings.css
@@ -1,12 +1,12 @@
-.executorChat-container {
+.expertRatings-container {
     width: fit-content;
     margin: 20px auto;
 }
-.executorChat-container button {
+.expertRatings-container button {
   margin: 20px auto;
   display: block;
 }
-.executorChat-container h1 {
+.expertRatings-container h1 {
   text-align: center;
 }
