+++ src/pages/Login/Login.jsx
@@ -1,5 +1,5 @@
 import { useNavigate } from 'react-router-dom';
-import './styles.css';
+import './Login.css';
 export const Login = () => {
   const navigate = useNavigate();
