+++ src/pages/ResetPassword/ResetPassword.jsx
@@ -1,5 +1,5 @@
 import { useNavigate } from 'react-router-dom';
-import './styles.css';
+import './ResetPassword.css';
 export const ResetPassword = () => {
   const navigate = useNavigate();
   return (
